# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","description":"YOLO-World API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","description":"Root endpoint with API information","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/model/info":{"get":{"summary":"Get Model Info","description":"Get current model configuration and status","operationId":"get_model_info_model_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelInfoResponse"}}}}}}},"/model/classes":{"get":{"summary":"Get Current Classes","description":"Get currently configured detection classes","operationId":"get_current_classes_model_classes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassesResponse"}}}}}},"post":{"summary":"Add Detection Classes","description":"Add new detection classes to the model","operationId":"add_detection_classes_model_classes_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassesRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Clear Detection Classes","description":"Clear all detection classes","operationId":"clear_detection_classes_model_classes_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/detect":{"post":{"summary":"Detect Object","description":"Detect objects in uploaded image","operationId":"detect_object_detect_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_detect_object_detect_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/detect/with-confidence":{"post":{"summary":"Detect Object With Confidence","description":"Detect objects in uploaded image with custom confidence threshold","operationId":"detect_object_with_confidence_detect_with_confidence_post","parameters":[{"name":"confidence","in":"query","required":false,"schema":{"type":"number","default":0.25,"title":"Confidence"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_detect_object_with_confidence_detect_with_confidence_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_detect_object_detect_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_detect_object_detect_post"},"Body_detect_object_with_confidence_detect_with_confidence_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_detect_object_with_confidence_detect_with_confidence_post"},"ClassesRequest":{"properties":{"classes":{"items":{"type":"string"},"type":"array","title":"Classes"}},"type":"object","required":["classes"],"title":"ClassesRequest"},"ClassesResponse":{"properties":{"classes":{"items":{"type":"string"},"type":"array","title":"Classes"},"message":{"type":"string","title":"Message"}},"type":"object","required":["classes","message"],"title":"ClassesResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"MessageResponse"},"ModelInfoResponse":{"properties":{"model_path":{"type":"string","title":"Model Path"},"vocab_file":{"type":"string","title":"Vocab File"},"current_classes":{"items":{"type":"string"},"type":"array","title":"Current Classes"},"total_classes":{"type":"integer","title":"Total Classes"}},"type":"object","required":["model_path","vocab_file","current_classes","total_classes"],"title":"ModelInfoResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}